# contracts/openapi/local-api.json
{
  "components": {
    "schemas": {
...
      "CsvMetadataImportBody": {
        "properties": {
          "csv_path": {
            "minLength": 1,
            "title": "Csv Path",
            "type": "string"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          }
        },
        "required": [
          "csv_path"
        ],
        "title": "CsvMetadataImportBody",
        "type": "object"
      },
      "CsvMetadataImportResponse": {
        "properties": {
          "ambiguous": {
            "title": "Ambiguous",
            "type": "integer"
          },
          "bpm_conflicts": {
            "title": "Bpm Conflicts",
            "type": "integer"
          },
          "csv_rows": {
            "title": "Csv Rows",
            "type": "integer"
          },
          "fields_enriched": {
            "title": "Fields Enriched",
            "type": "integer"
          },
          "key_conflicts": {
            "title": "Key Conflicts",
            "type": "integer"
          },
          "library_tracks": {
            "title": "Library Tracks",
            "type": "integer"
          },
          "matched": {
            "title": "Matched",
            "type": "integer"
          },
          "source_records_written": {
            "title": "Source Records Written",
            "type": "integer"
          },
          "source_ref": {
            "title": "Source Ref",
            "type": "string"
          },
          "tracks_enriched": {
            "title": "Tracks Enriched",
            "type": "integer"
          },
          "unmatched": {
            "title": "Unmatched",
            "type": "integer"
          }
        },
        "required": [
          "source_ref",
          "csv_rows",
          "library_tracks",
          "matched",
          "unmatched",
          "ambiguous",
          "source_records_written",
          "tracks_enriched",
          "fields_enriched",
          "bpm_conflicts",
          "key_conflicts"
        ],
        "title": "CsvMetadataImportResponse",
        "type": "object"
      },
...
      "FolderImportBody": {
        "properties": {
          "folder_path": {
            "minLength": 1,
            "title": "Folder Path",
            "type": "string"
          },
          "library_name": {
            "default": "Local Music",
            "title": "Library Name",
            "type": "string"
          }
        },
        "required": [
          "folder_path"
        ],
        "title": "FolderImportBody",
        "type": "object"
      },
      "FolderImportOutcome": {
        "additionalProperties": false,
        "properties": {
          "duplicate_of_track_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of Track Id"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "enum": [
              "imported",
              "duplicate",
              "unsupported",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "track_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "path",
          "status"
        ],
        "title": "FolderImportOutcome",
        "type": "object"
      },
      "FolderImportResponse": {
        "properties": {
          "examined": {
            "title": "Examined",
            "type": "integer"
          },
          "library_id": {
            "title": "Library Id",
            "type": "string"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/FolderImportOutcome"
            },
            "title": "Outcomes",
            "type": "array"
          },
          "scanned": {
            "title": "Scanned",
            "type": "integer"
          },
          "tracks": {
            "title": "Tracks",
            "type": "integer"
          }
        },
        "required": [
          "library_id",
          "scanned",
          "examined",
          "tracks",
          "outcomes"
        ],
        "title": "FolderImportResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TrackResponse": {
        "properties": {
          "album": {
            "title": "Album",
            "type": "string"
          },
          "artist": {
            "title": "Artist",
            "type": "string"
          },
          "audio_content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Content Hash"
          },
          "bpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm"
          },
          "bpm_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm Source"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "date_added": {
            "title": "Date Added",
            "type": "string"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "genre": {
            "title": "Genre",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "key_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Source"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "library_id": {
            "title": "Library Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "location_kind": {
            "title": "Location Kind",
            "type": "string"
          },
          "missing": {
            "title": "Missing",
            "type": "boolean"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "rekordbox_track_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rekordbox Track Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "library_id",
          "title",
          "artist",
          "album",
          "genre",
          "label",
          "bpm",
          "key",
          "duration_sec",
          "location",
          "location_kind",
          "audio_content_hash",
          "rating",
          "date_added",
          "rekordbox_track_id",
          "bpm_source",
          "key_source",
          "created_at",
          "missing",
          "preview_url"
        ],
        "title": "TrackResponse",
        "type": "object"
      },
...
      "TracksResponse": {
        "properties": {
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackResponse"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "tracks"
        ],
        "title": "TracksResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Crate Dig local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/audio/{track_id}": {
      "get": {
        "operationId": "audio_audio__track_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Stored local audio with HTTP Range support",
            "headers": {
              "Accept-Ranges": {
                "description": "Supported range unit",
                "schema": {
                  "const": "bytes",
                  "type": "string"
                }
              },
              "Content-Length": {
                "description": "Response body length in bytes",
                "schema": {
                  "type": "string"
                }
              },
              "Content-Type": {
                "description": "Detected audio media type",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "206": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Requested byte range of stored local audio",
            "headers": {
              "Accept-Ranges": {
                "description": "Supported range unit",
                "schema": {
                  "const": "bytes",
                  "type": "string"
                }
              },
              "Content-Length": {
                "description": "Partial response body length in bytes",
                "schema": {
                  "type": "string"
                }
              },
              "Content-Range": {
                "description": "Returned byte range and complete file size",
                "schema": {
                  "example": "bytes 0-31/1600",
                  "type": "string"
                }
              },
              "Content-Type": {
                "description": "Detected audio media type",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audio"
      }
    },
...
    "/imports/folder": {
      "post": {
        "operationId": "import_folder_imports_folder_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderImportBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Folder"
      }
    },
...
    "/libraries/{library_id}/metadata/import-csv": {
      "post": {
        "operationId": "import_csv_metadata_libraries__library_id__metadata_import_csv_post",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CsvMetadataImportBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsvMetadataImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Csv Metadata"
      }
    },
    "/libraries/{library_id}/tracks": {
      "get": {
        "operationId": "library_tracks_libraries__library_id__tracks_get",
        "parameters": [
          {
            "in": "path",
            "name": "library_id",
            "required": true,
            "schema": {
              "title": "Library Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Library Tracks"
      }
    },
...
    "/tracks/{track_id}": {
      "get": {
        "operationId": "one_track_tracks__track_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One Track"
      }
    },
...
  }
}